refactor fota code: Move code from example app into lib
