+++ example/lib/widgets/firmware_update_widget.dart
@@ -1,11 +1,10 @@
-import 'package:example/fota/src/bloc/bloc/update_bloc.dart';
-import 'package:example/fota/src/model/firmware_update_request.dart';
-import 'package:example/fota/src/providers/firmware_update_request_provider.dart';
-import 'package:example/fota/src/view/stepper_view/firmware_select.dart';
-import 'package:example/fota/src/view/stepper_view/update_view.dart';
+import 'package:open_earable_flutter/src/fota/bloc/bloc/update_bloc.dart';
+import 'package:open_earable_flutter/src/fota/model/firmware_update_request.dart';
+import 'package:open_earable_flutter/src/fota/providers/firmware_update_request_provider.dart';
+import 'package:open_earable_flutter/src/fota/view/stepper_view/firmware_select.dart';
+import 'package:open_earable_flutter/src/fota/view/stepper_view/update_view.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:open_earable_flutter/open_earable_flutter.dart';
 import 'package:provider/provider.dart';
 
 class FirmwareUpdateWidget extends StatefulWidget {
@@ -81,8 +80,6 @@ class _FirmwareUpdateWidget extends State<FirmwareUpdateWidget> {
           ],
         );
       case 1:
-        print("ID");
-        print(parameters.peripheral!.identifier);
         return BlocProvider(
           create: (context) => UpdateBloc(firmwareUpdateRequest: parameters),
           child: UpdateStepView(),
