+++ lib/src/fota/bloc/bloc/update_event.dart
