+++ lib/src/fota/bloc/bloc/update_state.dart
