+++ lib/src/fota/handlers/firmware_update_state.dart
