+++ lib/src/fota/model/firmware_image.dart
