+++ lib/src/fota/model/firmware_update_request.dart
@@ -1,7 +1,7 @@
 import 'dart:typed_data';
 
 import 'package:mcumgr_flutter/mcumgr_flutter.dart';
-import '../../src/model/firmware_image.dart';
+import '../model/firmware_image.dart';
 
 class FirmwareUpdateRequest {
   SelectedFirmware? firmware;
