+++ lib/src/fota/model/manifest.dart
