+++ lib/src/fota/utils/set_ext.dart
