+++ lib/src/fota/utils/string_ext.dart
