+++ lib/src/fota/view/logger_screen/logger_screen.dart
