+++ lib/src/fota/view/stepper_view/update_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import '../../../src/model/firmware_update_request.dart';
-import '../../../src/providers/firmware_update_request_provider.dart';
-import '../../../src/view/logger_screen/logger_screen.dart';
+import '../../model/firmware_update_request.dart';
+import '../../providers/firmware_update_request_provider.dart';
+import '../../view/logger_screen/logger_screen.dart';
 
 import '../../bloc/bloc/update_bloc.dart';
