+++ pubspec.yaml
@@ -23,6 +23,19 @@ dependencies:
   permission_handler: ^11.4.0
   universal_ble: ^0.17.0
   logger: ^2.5.0
+  # for fota:
+  equatable: ^2.0.5
+  file_picker: ^8.1.7
+  flutter_archive: ^6.0.0
+  flutter_bloc: ^8.1.3
+  http: ^1.1.2
+  json_annotation: ^4.8.1
+  mcumgr_flutter: ^0.4.2
+  path_provider: ^2.1.1
+  provider: ^6.1.1
+  rxdart: ^0.27.7
+  tuple: ^2.0.2
+  uuid: ^4.2.1
 
 dev_dependencies:
   flutter_test:
